% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remind_me_and_cheat.R
\name{cheat}
\alias{cheat}
\title{Return the first four solutions of Assignment 3.1}
\usage{
cheat(question_number_3.1)
}
\arguments{
\item{question_number_3.1}{numeric variable}
}
\description{
Return the first four solutions of Assignment 3.1
}
